Fixing empty callback bug
in newer versions of android sometimes without this check, none of the callbacks works. this is because of a new policy applied to package visibility.

From the documentation:
> When an app targets Android 11 (API level 30) or higher and queries
> for information about the other apps that are installed on a device,
> the system filters this information by default. The limited package
> visibility reduces the number of apps that appear to be installed on a
> device, from your app's perspective. The limited app visibility
> affects the return results of methods that give information about
> other apps, such as queryIntentActivities(), getPackageInfo(), and
> getInstalledApplications(). The limited visibility also affects
> explicit interactions with other apps, such as starting another app's
> service.

basically, the app can not see the bazaar and connect to its service. a simple workaround is to ask the user to open the cafe-bazaar app. 
i've also added 

    <queries>
            <package android:name="com.farsitel.bazaar" />
    </queries>

but it doesn't work. but for now, this will resolve the issue of empty callbacks